+++ NEWS.md
@@ -9,6 +9,7 @@
 - Subdaily soil temperature change limited for stability with thin first layer
 - Column 'Strict' added to SpParamsDefinition
 - Detection of POSIXct date columns
+- New control option 'lfmcComponent'
 
 # Version 4.3.1
 - Functions forest2spwbInput/forest2growthInput deprecated. Preferred function names are spwbInput/growthInput.